+++ openwa_bridge/hooks.py
@@ -218,7 +218,7 @@ override_doctype_class = {
 fixtures = [
 	{
 		"dt": "Custom Field",
-		"filters": [["dt", "in", ["WhatsApp Account", "WhatsApp Templates", "WhatsApp Notification"]]],
+		"filters": [["dt", "in", ["WhatsApp Account", "WhatsApp Templates", "WhatsApp Notification", "WhatsApp Message"]]],
 	},
 	{
 		"dt": "Property Setter",